requirements.txt
pyyaml==5.4.1
pyarrow==3.0.0
scikit-learn==0.24.1
pandas==1.2.2
google-cloud-bigquery==2.9.0
google-cloud-bigquery-storage==2.3.0
flask==1.1.2
Flask-RESTful==0.3.8
Flask-request-params==0.3.0
unidecode==1.2.0

# Documentation
sphinx_rtd_theme
sphinxemoji
nbsphinx==0.8.2
sphinx-copybutton==0.3.1